Replacing the Python InMemoryStore by the Rust version.
